by gacp

+++ Module-09/ex01/RPN.cpp
@@ -6,7 +6,7 @@
 /*   By: youmoukh <youmoukh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/11 15:31:30 by youmoukh          #+#    #+#             */
-/*   Updated: 2024/10/26 16:23:51 by youmoukh         ###   ########.fr       */
+/*   Updated: 2024/10/26 16:26:47 by youmoukh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -145,8 +145,8 @@ void	RPN::CheckFlow(int Value_1, int Value_2, char opr) const
 void	RPN::CalculateStaff(int opr, std::stack<float>& _stack)
 {
 	int	Y;
-	int	num1;
-	int	num2;
+	float	num1;
+	float	num2;
 	
 	// if operator then do simple math
